code/requirements.txt
about-time==3.1.1
aioboto3==10.4.0
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
alive-progress==2.4.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
boto3==1.24.59
botocore==1.27.59
certifi==2022.9.24
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1606236101993/work
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
decorator==4.4.2
deeplake==3.2.21
dill==0.3.6
efficientnet-pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastcluster==1.2.6
filetype==1.2.0
fonttools==4.28.5
frozenlist==1.3.3
future @ file:///home/conda/feedstock_root/build_artifacts/future_1649010148304/work
grapheme==0.6.0
humbug==0.3.1
idna==3.3
imageio==2.18.0
imageio-ffmpeg==0.4.7
ipython==8.1.1
jedi==0.18.1
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
moviepy==1.0.3
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==2.8
numcodecs==0.11.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1636145332760/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathos==0.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.11.0
plotly-express==0.4.1
pox==0.3.2
ppft==1.7.6.6
prettytable==3.7.0
proglog==0.1.10
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.11.2
PyJWT==2.6.0
pynvml==11.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
requests==2.27.1
s3transfer==0.6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SimpleITK==2.2.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.9.0
tenacity==8.1.0
tensorboardX==2.4.1
threadpoolctl==3.1.0
tifffile==2022.4.22
torch==1.10about-time==3.1.1
aioboto3==10.4.0
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
alive-progress==2.4.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
boto3==1.24.59
botocore==1.27.59
certifi==2022.9.24
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1606236101993/work
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
decorator==4.4.2
deeplake==3.2.21
dill==0.3.6
efficientnet-pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastcluster==1.2.6
filetype==1.2.0
fonttools==4.28.5
frozenlist==1.3.3
future @ file:///home/conda/feedstock_root/build_artifacts/future_1649010148304/work
grapheme==0.6.0
humbug==0.3.1
idna==3.3
imageio==2.18.0
imageio-ffmpeg==0.4.7
ipython==8.1.1
jedi==0.18.1
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
moviepy==1.0.3
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==2.8
numcodecs==0.11.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1636145332760/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathos==0.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.11.0
plotly-express==0.4.1
pox==0.3.2
ppft==1.7.6.6
prettytable==3.7.0
proglog==0.1.10
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.11.2
PyJWT==2.6.0
pynvml==11.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
requests==2.27.1
s3transfer==0.6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SimpleITK==2.2.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.9.0
tenacity==8.1.0
tensorboardX==2.4.1
threadpoolctl==3.1.0
tifffile==2022.4.22
torch==1.10.0+cu111
torch-cka==0.21
torchaudio==0.10.0+rocm4.1
torchsummary==1.5.1
torchvision==0.11.0+cu111
tqdm==4.63.1
traitlets==5.1.1
typing_extensions==4.5.0
unzip==1.0.0
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.15.0
yarl==1.8.2
zepid==0.9.1
